+++ backend/src/main/java/org/cathori/backend/alert/api/NotificationController.java
@@ -2,7 +2,7 @@ package org.cathori.backend.alert.api;
 
 import lombok.RequiredArgsConstructor;
 import org.cathori.backend.alert.api.dto.NotificationListResponse;
-import org.cathori.backend.alert.application.NotificationService;
+import org.cathori.backend.alert.application.notification.NotificationService;
 import org.cathori.backend.security.CustomUserDetails;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
