+++ backend/src/main/java/org/cathori/backend/alert/application/notification/NotificationQueryPort.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.alert.application;
+package org.cathori.backend.alert.application.notification;
 
 import java.util.List;
