+++ backend/src/main/java/org/cathori/backend/alert/application/notification/NotificationRow.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.alert.application;
+package org.cathori.backend.alert.application.notification;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
