+++ backend/src/main/java/org/cathori/backend/alert/application/push/PushNotificationPort.java
@@ -1,8 +1,8 @@
-package org.cathori.backend.alert.application;
+package org.cathori.backend.alert.application.push;
 
 import java.util.List;
 
-public interface FcmPort {
+public interface PushNotificationPort {
 
     List<UserFcmResult> sendMulticast(List<UserToken> targets, String title, String body, Long noticeId);
 }
