+++ backend/src/main/java/org/cathori/backend/alert/application/push/UserFcmResult.java
@@ -1,4 +1,4 @@
-package org.cathori.backend.alert.application;
+package org.cathori.backend.alert.application.push;
 
 public record UserFcmResult(Long userId, boolean success) {
 }
